+++ src/scanner/mod.rs
@@ -3,7 +3,9 @@ pub(crate) use enumerate::enumerate_filesystem_inputs;
 pub(crate) use repos::{clone_or_update_git_repos, enumerate_github_repos};
 pub use runner::{load_and_record_rules, run_async_scan, run_scan};
 pub(crate) use validation::run_secret_validation;
+pub(crate) use docker::save_docker_images;
 
+mod docker;
 mod enumerate;
 mod processing;
 mod repos;