+++ tests/int_github.rs
@@ -68,7 +68,9 @@ fn test_github_remote_scan() -> Result<()> {
 
             jira_url: None,
             jql: None,
-            max_results: 50,
+            max_results: 100,
+            // Docker image scanning
+            docker_image: Vec::new(),
             // git clone / history options
             git_clone: GitCloneMode::Bare,
             git_history: GitHistoryMode::Full,