+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.27.0]
 - Added Buildkite rule
+- Added support for scanning Docker images via `--docker-image`
 
 ## [1.26.0]
 - Added rule for ElevenLabs